change delta_x/y into double for more precision and better movement, also check collision after add x and y to move even if the player if against a wall

+++ include/struct.h
@@ -48,8 +48,8 @@ typedef struct s_player
 {
 	t_keys	*keys;
 	t_point	*pos;
-	int		delta_x;
-	int		delta_y;
+	double	delta_x;
+	double	delta_y;
 	double	angle;
 	double	fov;
 }			t_player;
